quorum: move lachowski to "his" own namespace
;)
